requirements.txt
# NBA Props Prediction Pipeline - Python Dependencies
# Updated for GitHub Actions deployment

# Core Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3

# NBA API
nba_api==1.4.1

# AWS S3
boto3==1.34.22
botocore==1.34.22

# Database (PostgreSQL/Neon)
psycopg2-binary==2.9.9

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0

# Utilities
python-dotenv==1.0.0

# Optional: Enhanced logging and monitoring
# colorlog==6.8.0
# sentry-sdk==1.39.2
